Implement Code-Quality
