+++ src/main/java/main/Launcher.java
@@ -1,5 +1,6 @@
 package main;
 
+import gui.Main;
 import javafx.application.Application;
 
 /**
@@ -7,6 +8,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Duke.class, args);
+        Application.launch(Main.class, args);
     }
 }
